Verification/requirements/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aiosqlite==0.20.0
amqp==5.3.1
ansi2html==1.9.2
anyio==4.6.2.post1
APScheduler==3.10.4
asgiref==3.8.1
asynctest==0.13.0
attrs==24.2.0
basedtyping==0.1.10
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.8
cron-descriptor==1.4.5
cryptography==43.0.3
Django==5.1.2
django-appconf==1.0.6
django-apscheduler==0.6.2
django-celery-beat==2.7.0
django-celery-email==3.0.0
django-jazzmin==3.0.1
django-oauth-toolkit==3.0.1
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-oauth==1.1.0
flower==2.0.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
importlib-resources==5.13.0
iniconfig==2.0.0
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kombu==5.4.2
MouseInfo==0.1.3
multidict==6.1.0
oauthlib==3.2.2
OSlash==0.6.3
packaging==24.1
pillow==11.1.0
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyotp==2.9.0
pyperclip==1.9.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
pyqt6rc==0.7.0
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
pytest==8.3.3
pytest-aiohttp==1.0.5
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-django==4.8.0
pytest-qt==4.4.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
qasync==0.27.1
qrcode==8.0
redis==5.2.0
referencing==0.35.1
requests==2.32.3
robotframework==7.1.1
robotframework-async==1.0.4
rpds-py==0.22.3
setuptools==75.2.0
shiboken6==6.8.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
yarl==1.18.3
